rename: 폴더 구조 개편, 리덕스 명칭 개선 feature/TODAK-172/auth-integration
