+++ app/(auth)/layout.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { Logo } from "@/components/layout";
-import useMocking from "@/hooks/useMocking";
-import store from "@/store";
+import store from "@/feature/redux";
+import { useMocking } from "@/utils/hooks";
 import { HR } from "flowbite-react";
 import { Provider } from "react-redux";
