+++ components/diary/Calendar.tsx
@@ -1,11 +1,11 @@
 "use client";
 
-import { RootState } from "@/store";
+import { RootState } from "@/feature/redux";
 import { useDispatch, useSelector } from "react-redux";
-import { setDate } from "@/store/slices/dairy/diarySlice";
+import { setDate } from "@/feature/redux/slices/dairy/diarySlice";
 import Calendar from "react-calendar";
 import { useRouter } from "next/navigation";
-import path from "@/routes";
+import path from "@/feature/routes";
 import "@/lib/react-calendar/Calendar.css";
 
 const MyCalendar: React.FC = () => {
