+++ feature/mocks/config/browser.ts
