+++ feature/mocks/domain/auth.ts
