+++ feature/mocks/domain/diary.ts
