+++ utils/hooks/index.ts
@@ -1,0 +1,4 @@
+import useHideOnScroll from "./useHideOnScroll";
+import useMocking from "./useMocking";
+
+export { useHideOnScroll, useMocking };
