+++ utils/types/animationType.ts
